CI test

+++ Makefile
@@ -22,33 +22,53 @@ all: generate docs
 # Build docker builder image
 .PHONY: build-image
 build-image:
+ifndef CIRCLECI
 	docker build -t $(BUILDIMAGE) -f Dockerfile.build .
+else
+	go get github.com/gogo/protobuf/protoc-gen-gogo && \
+    go get github.com/gogo/protobuf/protoc-gen-gofast && \
+    go get github.com/gogo/protobuf/protoc-gen-gogofaster && \
+    go get golang.org/x/tools/cmd/goimports && \
+    go get github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway && \
+    go get github.com/grpc-ecosystem/grpc-gateway/protoc-gen-swagger && \
+    go get github.com/golang/protobuf/protoc-gen-go && \
+    go get github.com/jessevdk/go-assets-builder && \
+    go get github.com/pseudomuto/protoc-gen-doc/cmd/protoc-gen-doc
+endif
 
 .PHONY: $(CACHEVOL)
 $(CACHEVOL):
+ifndef CIRCLECI
 	@docker volume create $(CACHEVOL)
 	docker run -it 	--rm -v $(CACHEVOL):/usr/gocache \
 		$(BUILDIMAGE) \
 		chown -R $(shell id -u):$(shell id -g) /usr/gocache
+endif
 
 .PHONY: $(MODVOL)
 $(MODVOL):
+ifndef CIRCLECI
 	@docker volume create $(MODVOL)
 	docker run -it 	--rm -v $(MODVOL):/go/pkg/mod \
 		$(BUILDIMAGE) \
 		chown -R $(shell id -u):$(shell id -g) /go/pkg/mod
+endif
 
 # Generate go code for proto files
 .PHONY: generate
 generate: $(CACHEVOL) $(MODVOL)
+ifndef CIRCLECI
 	docker $(DOCKERARGS) \
+endif
 		go generate ./...
 
 # Generate API docs
 .PHONY: docs
 docs: $(CACHEVOL) $(MODVOL)
 	@echo $(PROTOSOURCES)
+ifndef CIRCLECI
 	docker $(DOCKERARGS) \
+endif
 		protoc -I.:vendor:vendor/googleapis/:vendor/github.com/gogo/protobuf/protobuf/ \
 			--doc_out=docs $(PROTOSOURCES) \
 			--doc_opt=markdown,apis.md
